Replace internal data of FermionicOp
